requirements.txt
GitPython==3.1.43
pydantic==2.7.4
pydantic-settings==2.3.4
requests==2.32.3
